Chore (#80)
* Reverse CIplStore::GetIplEntityIndexArray

* Add CTempColModels::ms_colModelBBox

* Add CAudioZones::RegisterAudioBox and RegisterAudioSphere

* Rename VEHICLE_CLASS_NONE to VEHICLE_CLASS_IGNORE

* Reverse 2 cHandlingDataMgr functions

* Reverse CBaseModelInfo::ConvertAnimFileIndex

* Fix name CModelInfo::GetModelInfoUInt16

* Reverse LoadingScreen

* Fix COcclusion::AddOne signature

* Format CClumpModelInfo and CBaseModelInfo

* Add usage example cTransmission::DisplayGearRatios

* Add CBaseModelInfo::SetModelName

* Add new funcs to CEntryExitManager

* Use ReversibleHooks in CFileMgr

+++ source/game_sa/COcclusion.cpp
@@ -37,9 +37,10 @@ void COcclusion::Init()
     COcclusion::PreviousListWalkThroughFA = -1;
 }
 
-void COcclusion::AddOne(float dirMidX, float dirMidY, float dirMidZ, float widthX, float widthY, float height, float rotX, float rotY, float rotZ, float flags, bool isInterior)
+// 0x71DCD0
+void COcclusion::AddOne(float dirMidX, float dirMidY, float dirMidZ, float widthX, float widthY, float height, float rotX, float rotY, float rotZ, int flags, bool isInterior)
 {
-    plugin::Call<0x71DCD0, float, float, float, float, float, float, float, float, float, float, bool>
+    plugin::Call<0x71DCD0, float, float, float, float, float, float, float, float, float, int, bool>
         (dirMidX, dirMidY, dirMidZ, widthX, widthY, height, rotX, rotY, rotZ, flags, isInterior);
 }
